Proposing a first list of features
This is a first list of features. This is not a list I personally endorse, I tried to put here absolutely everything that was said on the mailing lists and on Gitter. We can later do some tidying a decide what we want to keep and what we want to throw away.